+++ test/check.js
@@ -1,9 +1,5 @@
 'use strict';
 
-/**
- * @deprecated Since version 3.0. Will be deleted in version 4.0. Use check instead.
- */
-
 var should = require('chai').should();
 
 describe('2valid tests', function () {
@@ -18,17 +14,15 @@ describe('2valid tests', function () {
 
     describe('Simple validate', function () {
 
-        console.warn("Calling deprecated function! Use check instead of validate");
-
         it('name', function (done) {
-            this.v.validate().should.eql({});
-            this.v.validate('some').should.eql({});
-            this.v.validate('string').should.eql({ notMatched: 'string' });
+            this.v.check().should.eql({});
+            this.v.check('some').should.eql({});
+            this.v.check('string').should.eql({ notMatched: 'string' });
             done();
           });
 
         it('object', function (done) {
-            this.v.validate({ id: { type: 'integer' } }).should.eql({});
+            this.v.check({ id: { type: 'integer' } }).should.eql({});
             done();
           });