+++ test/example.js
@@ -5,29 +5,29 @@ var should = require('chai').should();
 describe('example', function () {
 
     beforeEach(function () {
-        this.vm = require('../index');
+        this.v = require('../index');
       });
 
     afterEach(function () {
-        this.vm = null;
+        this.v = null;
       });
 
     describe('get example', function () {
 
         it('check types list', function (done) {
-            this.vm.getAllTypes().should.be.instanceof(Array);
-            this.vm.getAllTypes().should.contain('string');
-            this.vm.getAllTypes().should.contain('integer');
+            this.v.getAllTypes().should.be.instanceof(Array);
+            this.v.getAllTypes().should.contain('string');
+            this.v.getAllTypes().should.contain('integer');
             done();
           });
 
         it('check getExamples', function (done) {
-            var types = this.vm.getAllTypes();
+            var types = this.v.getAllTypes();
             for (var i = 0; i < types.length; i++) {
-              var example = this.vm.getExamples(types[i]);
+              var example = this.v.getExamples(types[i]);
               example.should.be.instanceof(Array);
               var opt = (types[i] !== 'any')? {} : {one:['Male','Female']};
-              this.vm.validate(types[i], example[0], opt, function (err) {
+              this.v.check(types[i], example[0], opt, function (err) {
                 (err === null).should.be.true;
               });
             }
@@ -36,11 +36,11 @@ describe('example', function () {
           });
 
         it('check all', function (done) {
-            var types = this.vm.getAllTypes();
+            var types = this.v.getAllTypes();
             for (var i = 0; i < types.length; i++) {
-              var example = this.vm.getExample(types[i]);
+              var example = this.v.getExample(types[i]);
               var opt = (types[i] !== 'any')? {} : {one:['Male','Female']};
-              this.vm.validate(types[i], example, opt, function (err) {
+              this.v.check(types[i], example, opt, function (err) {
                 (err === null).should.be.true;
               });
             }
@@ -49,11 +49,11 @@ describe('example', function () {
           });
 
         it('check all', function (done) {
-            var types = this.vm.getAllTypes();
+            var types = this.v.getAllTypes();
             for (var i = 0; i < types.length; i++) {
-              var example = this.vm.getRandomExample(types[i]);
+              var example = this.v.getRandomExample(types[i]);
               var opt = (types[i] !== 'any')? {} : {one:['Male','Female']};
-              this.vm.validate(types[i], example, opt, function (err) {
+              this.v.check(types[i], example, opt, function (err) {
                 (err === null).should.be.true;
               });
             }
@@ -62,12 +62,12 @@ describe('example', function () {
           });
 
         it('check notype', function (done) {
-            (this.vm.getExample('notype') === undefined).should.be.true;
+            (this.v.getExample('notype') === undefined).should.be.true;
             done();
           });
 
         it('check notype', function (done) {
-            (this.vm.getRandomExample('notype') === undefined).should.be.true;
+            (this.v.getRandomExample('notype') === undefined).should.be.true;
             done();
           });